drop generator-defaults
